Recomendation_API/requirements.txt
pydantic==2.8.2
fastapi==0.112.0
pydantic-settings==2.4.0
sqlalchemy==2.0.32
asyncpg==0.29.0
aiohttp==3.10.3
pytest-asyncio==0.23.8
python-jose==3.3.0
uvicorn==0.30.6
orjson==3.10.3
redis==4.4.2